Moved TempUser to its own file, changed URL to query param
to fix travis build error
query param:
?verify=CODEHERE
